v0.0.7 Client, v0.0.5: error handling implemented
