+++ functions/matchData.js
@@ -10,33 +10,38 @@ const cache = redis.createClient(process.env.REDIS_PORT);
 /*  Import helper function modules */
 const dynamoDb = require('./dynamoDbHelper');
 const keyBank = require('./cacheKeys');
+// Data Functions
+const Season = require('./seasonData');
+const Tournament = require('./tournamentData');
+const Profile = require('./profileData');
+const Team = require('./teamData');
 
 async function getMatchData(Id) {
     return new Promise(function(resolve, reject) {
         let cacheKey = keyBank.MATCH_PREFIX + Id;
         cache.get(cacheKey, async (err, data) => {
-            if (err) { reject(500); }
+            if (err) { console.error(err); reject(500); }
             else if (data != null) { resolve(JSON.parse(data)); }
             else {
                 try {
-                    let matchJson = await dynamoDb.getItem('Matches', 'MatchPId', req.params.matchId);
-                    if (matchJson == null) { reject(404); }
+                    let matchJson = await dynamoDb.getItem('Matches', 'MatchPId', Id);
+                    if (matchJson == null) { console.error("Match ID '" + Id + "' Not Found"); reject(404); }
                     else {
                         let seasonPId = matchJson['SeasonPId'];
-                        matchJson['SeasonShortName'] = await getSeasonShortName(seasonPId);
-                        matchJson['SeasonName'] = await getSeasonName(seasonPId);
+                        matchJson['SeasonShortName'] = await Season.getShortName(seasonPId);
+                        matchJson['SeasonName'] = await Season.getName(seasonPId);
                         let tourneyPId = matchJson['TournamentPId'];
-                        matchJson['TournamentShortName'] = await getTournamentShortName(tourneyPId);
-                        matchJson['TournamentName'] = await getTournamentName(tourneyPId);
+                        matchJson['TournamentShortName'] = await Tournament.getShortName(tourneyPId);
+                        matchJson['TournamentName'] = await Tournament.getName(tourneyPId);
                         let gameDurationMinute = matchJson['GameDuration'] / 60;
                         for (let i = 0; i < Object.keys(matchJson['Teams']).length; ++i) {
                             let teamId = Object.keys(matchJson['Teams'])[i];
                             let teamJson = matchJson['Teams'][teamId];
-                            teamJson['TeamName'] = await getTeamName(teamJson['TeamHId']);
+                            teamJson['TeamName'] = await Team.getName(teamJson['TeamHId']);
                             for (let j = 0; j < Object.keys(teamJson['Players']).length; ++j) {
                                 let partId = Object.keys(teamJson['Players'])[j];
                                 let playerJson = teamJson['Players'][partId];
-                                playerJson['ProfileName'] = await getProfileName(playerJson['ProfileHId']);
+                                playerJson['ProfileName'] = await Profile.getName(playerJson['ProfileHId']);
                                 playerJson['Kda'] = (playerJson['Deaths'] > 0) ? (((playerJson['Kills'] + playerJson['Assists']) / playerJson['Deaths']).toFixed(2)).toString() : "Perfect";
                                 playerJson['KillPct'] = ((playerJson['Kills'] + playerJson['Assists']) / teamJson['TeamKills']).toFixed(4);
                                 playerJson['DeathPct'] = (playerJson['Deaths'] / teamJson['TeamDeaths']).toFixed(4);
@@ -57,7 +62,7 @@ async function getMatchData(Id) {
                         resolve(matchJson);
                     }
                 }
-                catch (err) { reject(500); }
+                catch (err) { console.error(err); reject(500); }
             }
         });
     })
